Github tool succesfully added, prompts modified to work with new tool
