+++ common/prompts.py
@@ -1,71 +1,6 @@
-from langchain_core.prompts import PromptTemplate, ChatPromptTemplate, MessagesPlaceholder, HumanMessagePromptTemplate
-from typing import Any, Dict, List, Optional, Awaitable, Callable, Tuple, Type, Union
-import requests
-from datetime import datetime
-def fetch_github_events(username: str, target_date: str) -> str:
-    """
-    Fetch GitHub events for the user on the given date and return them as a human-readable string.
-    """
-    url = f"https://api.github.com/users/{username}/events/public"
-    response = requests.get(url)
-    if response.status_code != 200:
-        raise ValueError(f"Error: Received status code {response.status_code}")
-    events = response.json()
-    # Filter events by the target date
-    filtered_events = [
-        event for event in events
-        if datetime.fromisoformat(event['created_at'].replace("Z", "+00:00")).date() == datetime.strptime(target_date, "%Y-%m-%d").date()
-    ]
-
-    # Process events and extract details
-    event_details = []
-    for event in filtered_events:
-        event_type = event['type']
-        repo_name = event['repo']['name']
-        event_date = datetime.fromisoformat(event['created_at'].replace("Z", "+00:00")).strftime('%Y-%m-%d %H:%M:%S')  
-
-        details = f"Event Type: {event_type}\nRepository: {repo_name}\nDate: {event_date}"
-        # Event-specific details
-        if event_type == "PullRequestEvent":
-            pr_details = event['payload']['pull_request']
-            details += f"""
-Action: {event['payload']['action']}
-Pull Request URL: {pr_details['html_url']}
-Title: {pr_details['title']}
-Body: {pr_details['body']}"""
-        elif event_type == "PushEvent":
-            commits = event['payload']['commits']
-            commit_details = "\n".join(
-                f"  - Message: {commit['message']}\n    URL: {commit['url']}"
-                for commit in commits
-            )
-            details += f"\nCommits:\n{commit_details}"
-        elif event_type == "DeleteEvent":
-            details += f"""
-Ref Type: {event['payload']['ref_type']}
-Ref: {event['payload']['ref']}"""
-        elif event_type == "IssueCommentEvent":
-            comment_details = event['payload']['comment']
-            details += f"""
-Action: {event['payload']['action']}
-Issue URL: {event['payload']['issue']['html_url']}
-Comment: {comment_details['body']}"""
-        elif event_type == "IssuesEvent":
-            issue_details = event['payload']['issue']
-            details += f"""
-Action: {event['payload']['action']}
-Issue URL: {issue_details['html_url']}
-Title: {issue_details['title']}
-Body: {issue_details['body']}"""
-        elif event_type == "CreateEvent":
-            details += f"""
-Ref Type: {event['payload']['ref_type']}
-Ref: {event['payload'].get('ref', "N/A")}"""
-  
-        event_details.append(details)
-  
-    # Combine all events into a single formatted string
-    return "\n\n".join(event_details)
+from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
+from common.github_tool import fetch_github_events
+from datetime import datetime, timedelta
 
 ####### Welcome Message for the Bot Service #################
 WELCOME_MESSAGE = """
@@ -81,11 +16,17 @@ Feel free to ask any question and specify the tool you'd like me to utilize. I'm
 ---"""  
 ###########################################################  
   
-# Adding github stuff to the prompt  
-target_date_8 = "2025-01-08"
+# Get today's date, yesterday's date, and the day before yesterday's date
+today = datetime.now().date()
+yesterday = today - timedelta(days=1)
+day_before_yesterday = today - timedelta(days=2)
+
+# Format dates as strings
+target_date_8 = day_before_yesterday.strftime("%Y-%m-%d")
+target_date_9 = yesterday.strftime("%Y-%m-%d")
+target_date_today = today.strftime("%Y-%m-%d")
 username = "ashishkj21"
 GITHUB_INFO_8 = fetch_github_events(username, target_date_8)
-target_date_9 = "2025-01-9"
 GITHUB_INFO_9 = fetch_github_events(username, target_date_9)  
   
   
@@ -153,9 +94,9 @@ CUSTOM_CHATBOT_PREFIX = f"""
 - Answers from the tools are NOT considered part of the conversation. Treat tool's answers as context to respond to the human or to insert values into the database.  
 - Human does NOT have direct access to your tools.  
   
-## Github information needed to answer any question related to github. Github username is ashishkj21. Today is 9th January 2025. The github information from 8th January 2025 is:  
+## Github information needed to answer any question related to github. Github username is ashishkj21. Today is {target_date_today}. The github information from {target_date_8} is:  
 """ + GITHUB_INFO_8 + """  
-The github information from 9th January 2025 is:  
+The github information from {target_date_9} is:  
 """ + GITHUB_INFO_9 + """  
 Only refer to the above info for ANY question related to github.  
 """
