+++ common/utils.py
@@ -1,41 +1,13 @@
-import re
+from typing import Optional, Type
 import os
-import json
-from io import BytesIO
-from typing import Any, Dict, List, Optional, Awaitable, Callable, Tuple, Type, Union
-import asyncio
-
-from collections import OrderedDict
 from sqlalchemy.engine.url import URL
-
-from langchain.callbacks.manager import AsyncCallbackManagerForToolRun, CallbackManagerForToolRun
 from langchain.pydantic_v1 import BaseModel, Field, Extra
-from langchain.tools import BaseTool, StructuredTool, tool
-from typing import Dict, List
-from concurrent.futures import ThreadPoolExecutor
-from langchain.text_splitter import RecursiveCharacterTextSplitter
-from langchain.schema import BaseOutputParser, OutputParserException
-from langchain.chains import LLMChain
-from langchain.memory import ConversationBufferMemory
-from langchain.tools import BaseTool, StructuredTool, tool
-from langchain.prompts import PromptTemplate
+from langchain.tools import BaseTool
 from langchain.sql_database import SQLDatabase
-from langchain.agents import AgentExecutor, initialize_agent, AgentType, Tool
-from langchain.agents import create_sql_agent, create_openai_tools_agent
-from langchain_community.agent_toolkits import create_sql_agent
 from langchain_community.agent_toolkits import SQLDatabaseToolkit
-from langchain_community.utilities.sql_database import SQLDatabase
-from langchain.callbacks.base import BaseCallbackManager
-from langchain.requests import RequestsWrapper
-from langchain.chains import APIChain
-from langchain.agents.agent_toolkits.openapi.spec import reduce_openapi_spec
-from langchain.utils.json_schema import dereference_refs
+from langchain_community.agent_toolkits import create_sql_agent
 from langchain_openai import AzureChatOpenAI
-from langchain_openai import AzureOpenAIEmbeddings
-from langchain_core.prompts import ChatPromptTemplate
-from langchain_core.output_parsers import StrOutputParser
-from operator import itemgetter
-from typing import List
+from langchain.callbacks.manager import CallbackManagerForToolRun, AsyncCallbackManagerForToolRun
